+++ tests/integration/tools/mongodb/metadata/collectionSchema.test.ts
@@ -1,12 +1,10 @@
-import { describeWithMongoDB } from "../mongodbHelpers.js";
+import { describeWithMongoDB, validateAutoConnectBehavior } from "../mongodbHelpers.js";
 
 import {
     getResponseElements,
     getResponseContent,
-    setupIntegrationTest,
     dbOperationParameters,
     validateToolMetadata,
-    validateAutoConnectBehavior,
     validateThrowsForInvalidArguments,
     dbOperationInvalidArgTests,
 } from "../../../helpers.js";
