+++ tests/integration/tools/mongodb/metadata/collectionStorageSize.test.ts
@@ -1,11 +1,9 @@
-import { describeWithMongoDB } from "../mongodbHelpers.js";
+import { describeWithMongoDB, validateAutoConnectBehavior } from "../mongodbHelpers.js";
 
 import {
     getResponseContent,
-    setupIntegrationTest,
     dbOperationParameters,
     validateToolMetadata,
-    validateAutoConnectBehavior,
     dbOperationInvalidArgTests,
     validateThrowsForInvalidArguments,
 } from "../../../helpers.js";
